# Rename sparse contiguous() to coalesce(); make out of place; speed up THC Embedding 10x