--- a/torch/sparse/__init__.py
+++ b/torch/sparse/__init__.py
@@ -133,7 +133,6 @@ class _SparseBase(object):
         raise NotImplementedError
 
     def __str__(self):
-        self.coalesce_()  # to make sure the output is consistent
         return '{} with indices:\n{}and values:\n{}'.format(
             self.__class__.__name__, self.indices(), self.values())