… test/common.py
… test/test_sparse.py
… torch/lib/THCS/generic/THCSTensorMath.cu
… torch/lib/THPP/tensors/generic/THSTensor.cpp
… torch/lib/THS/generic/THSTensor.c
… torch/lib/THS/generic/THSTensor.h
… torch/lib/THS/generic/THSTensorMath.c
… torch/nn/_functions/thnn/sparse.py