Rename sparse tensor `contiguous` to `coalesce`, and be more sane about in-place vs out-of-place (i.e. always do it in place).

One thing I'm wondering: should the python methods `tensor.indices()`, `tensor.values()`, and `tensor.__repr__()` automatically coalesce first? This would have the advantage of hiding the implementation detail of coalescing from the Python user, guaranteeing expected invariants such as `t.indices() == t.coalesce_().indices()`. On the other hand, it limits what you can do in python (i.e. if you want to do something that uses the indices without coalescing for efficiency).

Opinions welcome!