@adamlerer 2017-04-19
Btw I'm still unsure about whether we should make `coalesce` an in-place op or not. It's an important decision and has some trade-offs

### Pro in-place:
* A coalesced tensor is always faster to operate on than an un-coalesced tensor.
* Suppose you have an uncoalesced sparse tensor `S` and you repeatedly call coalescing ops on it:
```
S = S + S # now uncoalesced
for i in range(N):
  T = S * S # coalesces each time
```
This is going to be extremely inefficient.

### Pro out-of-place:
* const operands should not be mutable; therefore, if a const operand (e.g. input to `cadd`) can get coalesced in place, it must be invisible to the Python user. That means that `t.indexes()`, `t.values()`, and `t.__str__` should all call coalesce. That hides opportunities for optimizations by munging `indices` and `values` directly from Python. Then if there are C bugs, you can get nasty behavior like a Heisenbug where when you call `print` on the tensor, the bug goes away.

### Always coalesce
There's a third option which is to always keep tensors in coalesced form. This is probably hard to make efficient. But it's possible that if we move `nnz` to be stored on the device, we can avoid host syncs and make it reasonably fast.