@ezyang 2017-04-24
I'm not too qualified to review this diff, but some thoughts did occur to me:

1. Does it make sense to still offer an in-place coalesce, even though operators won't call this directly? The main use case would be a user is performing a bunch of sparse ops in place, and then wants to finally coalesce the tensor before passing it off.
2. It would be great if we had some docs, because there is quite a bit of diversity in how duplicate indexes are handled. SciPy, for example, sums duplicate indexes together (like us, if I understand these diffs correctly), but TensorFlow drops duplicates (https://github.com/tensorflow/tensorflow/issues/371). I know the API is in flux, but it would still be great to have a paragraph or two stating what the basic assumptions in the implementation are.