@adamlerer 2017-04-24
Thanks @ezyang
 
1. I don't see what value an in-place op provides... the in-place coalesce is no more efficient than out-of place, because a sparse tensor is just a tiny wrapper for {THLongTensor indices, THTensor values}, so all you're creating is a new wrapper. The difference for the in-place one is that if anyone else was holding a reference to the tensor, it would be changed as well. Is there a reason we want that?

2. Yeah, we definitely need docs. I will at least put a comment in `coalesce` but we do need to write the full docs for this library soon. I won't be able to get to it until at least late-May, so if you'd like to pick this up, it would be a huge help.

Re: duplicates, the current implementation must sum duplicates because we take advantage of this to do some ops efficiently. `cadd` just "adds" two sparse tensors by concatenating their indices and values. This way you can add together a bunch of tensors and then accumulate them into a dense tensor, without ever doing a (slow) `coalesce`. 
P.S. We may be able to make `coalesce` much faster and then wouldn't need this, but that would require some substantial and tricky rewriting (esp. holding `nnz` on the GPU to avoid a host sync after `coalesce`).