@adamlerer 2017-04-24
Note to self: in the torch tests, `self.assertEqual` calls `x.coalesce()` on sparse tensors before comparison. But if `coalesce` is broken (e.g. to always return an empty tensor) these tests will pass. 
If it's fast enough, I want to write a simple Python coalesce, and then replace `x.coalesce()` with `checkedCoalesce(x)`, which runs the Python and torch versions and asserts they're equal.