@adamlerer 2017-04-27
This is ready for review and merge. @apaszke you'll note that you were right in #1147 to suggest making coalesce out-of-place.

Re: Embedding perf, here's the results of the benchmark script at https://gist.github.com/adamlerer/4159ddb4bd19e87de1b4bb3215a1db6f

```
$ python benchmark_sparse.py  1
cuda=True
Old done in 0.493923 s
New done in 0.617863 s
New done in 0.382167 s (no optimizer)
```

cf. 3s for new before this PR.